:sparkles: snapshotSandbox for IE compatible
