+++ src/interfaces.ts
@@ -36,3 +36,16 @@ export type StartOpts = {
 
 export type Rebuilder = () => void;
 export type Freer = () => Rebuilder;
+
+export interface SandBox {
+  /** 沙箱的名字 */
+  name: string;
+  /** 沙箱导出的代理实体 */
+  proxy: WindowProxy;
+  /** 沙箱是否在运行中 */
+  sandboxRunning: boolean;
+  /** 启动沙箱 */
+  active(): void;
+  /** 关闭沙箱 */
+  inactive(): void;
+}
